requirements.txt
# pycorenlp
lark-parser